Added and fixed Javadocs
