+++ qa/integration/src/test/java/org/eclipse/kapua/integration/service/device/management/RunDeviceManagementKeystoreI9nTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2021 Eurotech and/or its affiliates and others
+ * Copyright (c) 2021 Eurotech and/or its affiliates and others
  *
  * This program and the accompanying materials are made
  * available under the terms of the Eclipse Public License 2.0
@@ -21,10 +21,11 @@ import org.junit.runner.RunWith;
         features = {
                 "classpath:features/deviceManagement/DeviceManagementKeystoreI9n.feature",
         },
-        glue = {"org.eclipse.kapua.service.user.steps",
+        glue = {
                 "org.eclipse.kapua.qa.common",
                 "org.eclipse.kapua.service.account.steps",
                 "org.eclipse.kapua.service.device.registry.steps",
+                "org.eclipse.kapua.service.user.steps",
         },
         plugin = {"pretty",
                 "html:target/cucumber/DeviceManagementKeystoreI9n",
