+++ service/device/management/keystore/internal/src/main/java/org/eclipse/kapua/service/device/management/keystore/model/internal/DeviceKeystoreCertificateImpl.java
@@ -14,12 +14,22 @@ package org.eclipse.kapua.service.device.management.keystore.model.internal;
 
 import org.eclipse.kapua.service.device.management.keystore.model.DeviceKeystoreCertificate;
 
+/**
+ * {@link DeviceKeystoreCertificate} implementation.
+ *
+ * @since 1.5.0
+ */
 public class DeviceKeystoreCertificateImpl implements DeviceKeystoreCertificate {
 
     private String keystoreId;
     private String alias;
     private String certificate;
 
+    /**
+     * Constructor.
+     *
+     * @since 1.5.0
+     */
     public DeviceKeystoreCertificateImpl() {
     }
