+++ service/device/management/keystore/internal/src/main/java/org/eclipse/kapua/service/device/management/keystore/model/internal/DeviceKeystoreItemsImpl.java
@@ -18,12 +18,22 @@ import org.eclipse.kapua.service.device.management.keystore.model.DeviceKeystore
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * {@link DeviceKeystoreItems} implementation.
+ *
+ * @since 1.5.0
+ */
 public class DeviceKeystoreItemsImpl implements DeviceKeystoreItems {
 
     private static final long serialVersionUID = 1318691706440015345L;
 
     private List<DeviceKeystoreItem> keystoreItems;
 
+    /**
+     * Constructor.
+     *
+     * @since 1.5.0
+     */
     public DeviceKeystoreItemsImpl() {
     }
