+++ translator/kapua/kura/src/main/java/org/eclipse/kapua/translator/kura/kapua/keystore/TranslatorAppKeystoresKuraKapua.java
@@ -17,10 +17,11 @@ import org.eclipse.kapua.service.device.call.message.kura.app.response.KuraRespo
 import org.eclipse.kapua.service.device.call.message.kura.app.response.KuraResponsePayload;
 import org.eclipse.kapua.service.device.management.keystore.message.internal.response.KeystoreResponsePayload;
 import org.eclipse.kapua.service.device.management.keystore.message.internal.response.KeystoresResponseMessage;
+import org.eclipse.kapua.translator.Translator;
 import org.eclipse.kapua.translator.exception.InvalidPayloadException;
 
 /**
- * {@link org.eclipse.kapua.translator.Translator} implementation from {@link KuraResponseMessage} to {@link KeystoresResponseMessage}
+ * {@link Translator} implementation from {@link KuraResponseMessage} to {@link KeystoresResponseMessage}
  *
  * @since 1.5.0
  */
